refactor(solver): split rest arity helpers
