+++ crates/tsz-solver/src/operations/call_args.rs
@@ -16,8 +16,6 @@ use rustc_hash::{FxHashMap, FxHashSet};
 use std::cell::RefCell;
 use tracing::trace;
 
-const SPREAD_ARGUMENT_MARKER_NAME: &str = "__tsz_spread_argument__";
-
 // Reusable scratch `FxHashSet<crate::TypeId>` for the recursive DFS used by
 // `type_evaluates_to_function`. Mirrors the pool pattern from #4722 / #4790
 // / #4801 / #4805.
@@ -928,37 +926,6 @@ impl<'a, C: AssignabilityChecker> CallEvaluator<'a, C> {
             })
             .collect();
         self.interner.tuple(elements)
-    }
-
-    pub(crate) fn spread_argument_marker_inner(&self, type_id: TypeId) -> Option<TypeId> {
-        let Some(TypeData::Tuple(elems_id)) = self.interner.lookup(type_id) else {
-            return None;
-        };
-        let elems = self.interner.tuple_list(elems_id);
-        let [elem] = &*elems else {
-            return None;
-        };
-        if !elem.rest {
-            return None;
-        }
-        let name = elem.name?;
-        (self.interner.resolve_atom(name) == SPREAD_ARGUMENT_MARKER_NAME).then_some(elem.type_id)
-    }
-
-    fn generic_spread_argument_marker_inner(&self, type_id: TypeId) -> Option<TypeId> {
-        let Some(TypeData::Tuple(elems_id)) = self.interner.lookup(type_id) else {
-            return None;
-        };
-        let [elem] = &*self.interner.tuple_list(elems_id) else {
-            return None;
-        };
-        (elem.rest
-            && elem.name.is_none()
-            && matches!(
-                self.interner.lookup(elem.type_id),
-                Some(TypeData::TypeParameter(_))
-            ))
-        .then_some(elem.type_id)
     }
 
     fn normalize_spread_actual_type(&mut self, type_id: TypeId) -> TypeId {
