+++ crates/tsz-solver/src/operations/mod.rs
@@ -50,6 +50,7 @@ pub mod property;
 mod property_readonly;
 mod property_visitor;
 pub(crate) mod sequence_property;
+mod spread_args;
 pub mod widening;
 
 // Re-exports from core implementation
